main/poppler: move qt into main package, enable gpgme
depending on gpgme doesn't seem to introduce a cycle anymore

the libjpeg-turbo-devel makedep was added as well since it was in the
separate -qt template
