+++ main/libpoppler-qt6
@@ -1,0 +1,1 @@
+poppler
