--- main/poppler-qt-devel
@@ -1,1 +1,0 @@
-poppler-qt
