--- main/poppler-qt/patches
@@ -1,1 +1,0 @@
-../../main/poppler/patches
