--- main/poppler-qt/update.py
@@ -1,1 +1,0 @@
-../../main/poppler/update.py
